fix(ProgressBar): updates styles
